+++ bpf/sched_ext/samples/hello.bpf.c
@@ -1,0 +1,7 @@
+#include <vmlinux.h>
+#include <bpf/bpf_helpers.h>
+#include <bpf/bpf_endian.h>
+#include <bpf/bpf_tracing.h>
+#include <bpf/bpf_core_read.h>
+
+char __license[] SEC("license") = "GPL";